djangotutorial/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.13.0
whitenoise==6.9.0
